Atualizações em Controllers e em variáveis
AssignmentController:

URL "assignments" mudou para "request-assignment"
implementação função finish assignment no caminho "/finish-assignment/{id}"

UserControler:

antigo GetMapping("/login-user") mudou para um método post
implementação da função listAssignmentsInProgress no caminho "/assingments-in-progress/{id_user}"

User(model):

average_rating inicia como 5F
services_count inicia como 0

Assignment(model):
Double payment --> Float payment

Todos:
Mudança de método de criação de id(Autoincrement para UUID) e adaptações em todas as classes e funções relacionadas

extras:
assignmentForLibrary: Double payment--> Float payment


Continuidade: Filtro

+++ src/main/java/br/com/philance/backend/DTO/response/assignment/AssignmentForLibraryDTO.java
@@ -7,7 +7,7 @@ public record AssignmentForLibraryDTO(
         String company,
         String address,
         String description,
-        Double payment,
+        Float payment,
         int min_age,
         String attire
 ) {
